requirements.txt
celery==5.4.0
cookiecutter==2.6.0
django-browser-reload==1.17.0
django-tailwind==3.8.0
Django==5.1.6
djlint==1.36.4
gunicorn==23.0.0
pip-tools==7.4.1  
pipdeptree==2.25.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
phonenumbers==8.13.31
django-phonenumber-field==6.0.0
google-genai==1.24.0
sentry-sdk==2.32.0
redis>=4.5.0
django-redis>=5.2.0
google-genai==1.24.0
sentry-sdk==2.32.0
redis>=4.5.0
django-redis>=5.2.0
whitenoise==6.9.0